Average time in seconds
Now the time profiling presents the results in seconds, and in the organized manner  that we were asked to deliver the results for data structures.
At this point we can consider the profiler done, even if there are things that we can add to the system, but the main functionality is here and ready to work.